Use bootstrap 3 modal for service configuration
I don't know why the 'visible' system is used, by .modal() is never called anywhere, so I have changed this.
To launch service configuration, we use $(element).modal() now, and to hide it $(element).modal('hide').
I've added a class="modal" <div/> around the configurationServiceLoginDialog so that it works correctly, and is according to Bootstrap's own documentation for modals.
Added set ft=handlebars to some .html files so that they use the right syntax in vim (convenient for vim users)